Update Hitdie_Control.js
So subtracting a hit die can set your hit dice to zero. For a quick repro, try subtracting one hit die twice - you should end up with 0.

I took off the clamp to see what was happening and it seems like instead of an addition operation, it's doing a string concatenation - you start with 0, subtract 1HD: 01, subtract 1HD: 011 - which, if you're below level 11, clamps to your character level.

I glued in some parseInt operations when grabbing the value and it seems to be functional again. There's probably a more elegant way to do it but I don't know any javascript.